Simplify, refactor transformations to share more code.
